✓ mon-pix: fix user certification details integration tests
