+++ mon-pix/app/styles/components/_user-certifications-detail-competence.scss
@@ -119,8 +119,4 @@
   &--disabled {
     opacity: 0.4;
   }
-
-  &__level--disabled {
-    display: none;
-  }
 }
