+++ mon-pix/tests/integration/components/user-certifications-detail-competence_test.js
@@ -2,19 +2,16 @@ import { module, test } from 'qunit';
 import setupIntlRenderingTest from '../../helpers/setup-intl-rendering';
 import { hbs } from 'ember-cli-htmlbars';
 import EmberObject from '@ember/object';
-// eslint-disable-next-line no-restricted-imports
-import { find } from '@ember/test-helpers';
 import { render } from '@1024pix/ember-testing-library';
 import { A as EmberArray } from '@ember/array';
+import { getTableData } from '../../helpers/get-table-data';
 
 module('Integration | Component | user-certifications-detail-competence', function (hooks) {
   setupIntlRenderingTest(hooks);
 
   let area;
+  let screen;
   const PARENT_SELECTOR = '.user-certifications-detail-competence';
-  const TITLE_SELECTOR = `${PARENT_SELECTOR}__title`;
-  const COMPETENCE_SELECTOR = `${PARENT_SELECTOR} div`;
-  const DISABLED_CLASS = 'user-certifications-detail-competence__competence--disabled';
 
   hooks.beforeEach(async function () {
     area = EmberObject.create({
@@ -52,7 +49,7 @@ module('Integration | Component | user-certifications-detail-competence', functi
     this.set('area', area);
 
     // when
-    await render(hbs`<UserCertificationsDetailCompetence @area={{this.area}} />`);
+    screen = await render(hbs`<UserCertificationsDetailCompetence @area={{this.area}} />`);
   });
 
   test('renders', function (assert) {
@@ -60,39 +57,36 @@ module('Integration | Component | user-certifications-detail-competence', functi
     assert.dom(`${PARENT_SELECTOR}`).exists();
   });
 
-  test('should show the title of area', function (assert) {
+  test('should show the title of area', async function (assert) {
     // then
-    assert.ok(find(TITLE_SELECTOR).textContent.includes(area.title));
+    const title = await screen.getByRole('columnheader', {
+      name: area.title,
+    });
+    assert.dom(title).exists();
   });
 
-  test('should include one competences line per competence', function (assert) {
+  test('render all the competences properly', async function (assert) {
     // then
-    assert.dom(COMPETENCE_SELECTOR).exists({ count: area.resultCompetences.length });
-  });
+    const table = await screen.getByRole('table');
 
-  module('on a specific line of competence', function () {
-    module('when competence level is -1', function () {
-      test('should be grayed out (almost transparent) and not show the level', function (assert) {
-        assert.strictEqual(
-          find(`${COMPETENCE_SELECTOR}:nth-child(3) span`).textContent,
-          area.resultCompetences[1].level.toString(),
-        );
-        assert.ok(find(`${COMPETENCE_SELECTOR}:nth-child(3) p`).classList.toString().includes(DISABLED_CLASS));
-      });
-    });
-
-    module('when competence level is 0', function () {
-      test('should show "-" for the level (not 0)', function (assert) {
-        assert.strictEqual(find(`${COMPETENCE_SELECTOR}:nth-child(4) span`).textContent, '-');
-        assert.notOk(find(`${COMPETENCE_SELECTOR}:nth-child(4) p`).classList.toString().includes(DISABLED_CLASS));
-      });
-    });
-
-    module('when competence level is greater or equal than 1', function () {
-      test('should show the level', function (assert) {
-        assert.strictEqual(find(`${COMPETENCE_SELECTOR}:nth-child(5) span`).textContent, '3');
-        assert.notOk(find(`${COMPETENCE_SELECTOR}:nth-child(5) p`).classList.toString().includes(DISABLED_CLASS));
-      });
-    });
+    const tableData = await getTableData(table);
+    assert.deepEqual(tableData, [
+      {
+        'CRÉATION DE CONTENU': 'Programmer',
+        NIVEAU: '5',
+      },
+      {
+        'CRÉATION DE CONTENU': 'Adapter les docs à leur finalité',
+        NIVEAU: '',
+      },
+      {
+        'CRÉATION DE CONTENU': 'Développer des docs multimédia',
+        NIVEAU: '',
+      },
+      {
+        'CRÉATION DE CONTENU': 'Développer des docs textuels',
+        NIVEAU: '3',
+      },
+    ]);
   });
 });
